Move setup.d scripts to separate directory once finished
This change enables continuing the execution of setup.sh after an error has occurred. This makes handling errors significantly easier.

The person performing the setup still needs to clean up and (re)move the failed setup.d script manually, but once that is done, a second call to setup.sh will simply execute those scripts that still remain in setup.d. A problem that remains, however, is the passing of variables between setup.d scripts. For example, if 01-read-pw.sh assigns a value to $PASSWORD and execution of 02-whatever.sh fails, then 03-use-pw.sh will not have the password available to it.