Update postfix virtual alias database example
Some software (e. g. MariaDB < 10.0.1) has trouble with "DEFAULT CURRENT_TIMESTAMP" in this configuration. In such cases, the workaround is to use another suitable default and have the user insert the current date for `created` manually, which will then automatically change the `updated` timestamp.

The `created` field is changed to a simple date to accommodate this, and it is also quite sufficient in practice.

  `created` DATE DEFAULT NULL,
  `updated` TIMESTAMP NOT NULL DEFAULT '0000-00-00 00:00:00' ON UPDATE CURRENT_TIMESTAMP,